fix:removed logs and refactored tts
